% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundancy.R
\name{estimateRedundancy}
\alias{estimateRedundancy}
\title{Estimates the redundancy of enriched gene sets}
\usage{
estimateRedundancy(scores, gsets, Gdiffexp)
}
\arguments{
\item{scores}{Dataframe from `r packagedocs::rd_link(testUnidirectionality())` containing the filtered enrichment scores for every gene set}

\item{gsets}{Named list containing character vectors for each gene set}

\item{Gdiffexp}{Character vector of differentially expressed genes}
}
\value{
Numeric vector containing for every gene set (in the same order the scores dataframe) its relevance compared with other gene sets, higher is better
}
\description{
Wrapper around the `mgsa` method
}

