% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getTransformationMatrix}
\alias{getTransformationMatrix}
\title{Barycentric transformation matrix}
\usage{
getTransformationMatrix(anglebase = 0)
}
\arguments{
\item{anglebase}{Number of radians the first barycentric direction should be rotated anticlockwise}
}
\description{
Get the matrix for the barycentric transformation
}

