% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{interactiveDotplot}
\alias{interactiveDotplot}
\title{Interactive triwise dotplot}
\usage{
interactiveDotplot(Eoi, Gdiffexp = rownames(Eoi), Goi = c(),
  Glabels = rownames(Eoi), Gpin = c(), Coi = colnames(Eoi),
  colorvalues = NULL, rmax = 5, sizevalues = c(T = 2, F = 0.5),
  alphavalues = c(T = 0.8, F = 0.8), plotLocalEnrichment = F,
  width = NULL, height = NULL)
}
\arguments{
\item{Eoi}{Expression matrix with the three conditions in the columns}

\item{Gdiffexp}{Differentially expressed genes}

\item{Goi}{Genes of interest, a character or numeric vector to plot one set of genes, a named list containing different such vectors to plot multiple gene sets}

\item{Glabels}{Labels for every gene if different from the rownames of `Eoi`}

\item{Gpin}{Pinned genes, if `NULL` will automatically choose the top 20 most differentially expressed genes}

\item{Coi}{Character vector specifying the names of the three biological conditions, used for labelling}

\item{colorvalues}{Color values, different syntax depending on the colorby parameter: \itemize{
  \item diffexp: Named list with colors. First part of the name denotes whether a gene is differentially expressed (`diffexp` or `nodiffexp`). The second part denotes the name of the gene set. Genes not within a gene set are denoted by `all`. \cr
  For example: list(diffexpall="#000000", nodiffexpall="#AAAAAA", nodiffexpgset="#FFAAAA", diffexpgset="#FF0000")
  \item z: A character vector of color values, used to generate the color gradient
}}

\item{rmax}{Number denoting the maximal radius of the grid. All points outside of the grid will be clipped on the boundaries.}

\item{sizevalues}{Named list with the size of each dot if differentially expressed (TRUE) or not (FALSE)}

\item{alphavalues}{Named list with the alpha value of each dot if differentially expressed (TRUE) or not (FALSE)}

\item{plotLocalEnrichment}{Whether to plot local enrichment as a ring around the dotplot}

\item{width}{Width of the plot}

\item{height}{Height of the plot}
}
\description{
Draw an interactive triwise dotplot using a html widget
}
\examples{
data(vandelaar)
Eoi <- limma::avearrays(vandelaar, Biobase::phenoData(vandelaar)$celltype)
Eoi = Eoi[,c("YS_MF", "FL_mono", "BM_mono")]
barycoords = transformBarycentric(Eoi)
\dontrun{
interactiveDotplot(barycoords)
}

Eoi = matrix(rnorm(1000*3, sd=0.5), 1000, 3, dimnames=list(1:1000, c(1,2,3)))
Eoi[1:100,1] = Eoi[1:100,1] + 1
barycoords = transformBarycentric(Eoi)
Gdiffexp =(1:1000)[barycoords$r > 1]

\dontrun{
interactiveDotplot(Eoi)
interactiveDotplot(Eoi, Gdiffexp)
interactiveDotplot(Eoi, as.character(Gdiffexp), as.character(1:10), as.character(1:1000))
interactiveDotplot(Eoi, as.character(Gdiffexp), as.character(1:10), as.character(1:1000), c(50, 200))
}
}

