% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{named.list}
\alias{named.list}
\title{make a named list using the variable names (avoids repeating the same name twice)}
\usage{
named.list(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
make a named list using the variable names (avoids repeating the same name twice)
}

