% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plotDotplot}
\alias{plotDotplot}
\title{Creating a dotplot}
\usage{
plotDotplot(barycoords, diffexp = NULL, gsets = NULL,
  conditions = attr(barycoords, "conditions"), colorby = "diffexp",
  colors = NULL, rmax = 5)
}
\arguments{
\item{barycoords}{Dataframe containing for every gene its barycentric coordinates, as return by \code{transformBarycentric}}

\item{gsets}{Genes of interest, a character or numeric vector to plot one set of genes, a named list to plot multiple gene lists}

\item{conditions}{Character vector specifying the names of the three biological conditions}

\item{colorby}{Color by differential expression ("diffexp") or by log fold-change ("z")}

\item{Gdiffexp}{Differentially expressed genes}
}
\value{
Dataframe containing for every original point its new x and y coordinates in barycentric space
}
\description{
Plot a dotplot
}

