% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{regularsimplex}
\alias{regularsimplex}
\title{Coordinates of a regular simplex}
\usage{
regularsimplex(d)
}
\arguments{
\item{E}{Expression matrix}
}
\value{
Dataframe containing for every original point its new x and y coordinates in barycentric space
}
\description{
Get the coordinates of a regular simplex with `d+1` corners in `d` dimensions
}

