% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze2.R, R/analyze3.R
\name{testEnrichment}
\alias{testEnrichment}
\title{Simple test for enrichment in a given set of genes of interest (Goi)}
\usage{
testEnrichment(Goi, gsets, background, minknown = 2, minfound = 2,
  maxknown = 500)

testEnrichment(Goi, gsets, background, minknown = 2, minfound = 2,
  maxknown = 500)
}
\arguments{
\item{Goi}{character vector containing the genes tested for enrichment}

\item{gsets}{a GeneSetCollection or a}

\item{background}{character vector}

\item{Goi}{character vector containing the genes tested for enrichment}

\item{gsets}{a GeneSetCollection or a}

\item{background}{character vector}
}
\value{
Dataframe containing for every gene set which passed the filtering: \itemize{
 \item p-value of enrichment
 \item q-value of enrichment (p-value corrected for multiple testing)
 \item estimated odds score of enrichment, how much more likely is a given gene set to be part of
}

Dataframe containing for every gene set which passed the filtering: \itemize{
 \item p-value of enrichment
 \item q-value of enrichment (p-value corrected for multiple testing)
 \item estimated odds score of enrichment, how much more likely is a given gene set to be part of
}
}
\description{
Simple test for enrichment in a given set of genes of interest (Goi)

Simple test for enrichment in a given set of genes of interest (Goi)
}

