% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{testLocality}
\alias{testLocality}
\title{Test Locality
Test local upregulation}
\usage{
testLocality(Goi, Gdiffexp, angles, deltangle = pi/24, bandwidth = pi/3)
}
\description{
Test Locality
Test local upregulation
}

