% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{testUnidirectionality}
\alias{testUnidirectionality}
\title{Test gene sets for unidirectional enrichment}
\usage{
testUnidirectionality(barycoords, gsets, Gdiffexp = NULL,
  statistic = "diffexp", bm = NULL, minknown = 5, mindiffexp = 5,
  maxknown = 1500, mc.cores = getOption("mc.cores", default = 1))
}
\arguments{
\item{gsets}{list of character vectors, each containing a set of genes (gene identifiers)}

\item{Gdiffexp}{differentially expressed genes}

\item{angles}{either a dataframe obtained from transformBarycentric or a list of angles with gene identifiers as names}
}
\value{
Dataframe containing for every gene set which passed the filtering: \itemize{
 \item p-value of unidirectionality
 \item q-value of unidirectionality (p-value corrected for multiple testing)
 \item average angle
}
}
\description{
Test gene sets for unidirectional enrichment
}

