% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{transformReverseBarycentric}
\alias{transformReverseBarycentric}
\title{Transform barycentric coordinates in `d-1` dimensions back to regular coordinates in `d` dimensions (centered around 0)}
\usage{
transformReverseBarycentric(barypoints)
}
\arguments{
\item{barypoints}{Barycentric coordinates}
}
\value{
E Expression matrix centered around zero
}
\description{
Transform barycentric coordinates in `d-1` dimensions back to regular coordinates in `d` dimensions (centered around 0)
}

