% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{generateBackgroundModel}
\alias{generateBackgroundModel}
\title{Generate background models}
\usage{
generateBackgroundModel(barycoords, noi = seq(5, 100, 5),
  anglesoi = seqClosed(0, 2 * pi, 24), nsamples = 1e+05, bw = 20,
  mc.cores = getOption("mc.cores", default = 1))
}
\description{
Generates a background model by randomly resampling genes at different number of genes and calculating z distributions at different fixed angles
}

