% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{interactiveDotplot}
\alias{interactiveDotplot}
\title{Interactive triwise dotplot}
\usage{
interactiveDotplot(Eoi, Gdiffexp = c(), Goi = c(),
  Glabels = rownames(Eoi), Gpin = c(), plotLocalEnrichment = T,
  width = NULL, height = NULL)
}
\arguments{
\item{Eoi}{Expression matrix with the three conditions in the columns}

\item{Gdiffexp}{Differentially expressed genes}

\item{Goi}{List with genes of interest}

\item{Glabels}{Labels for every gene if different from the rownames of `Eoi`}

\item{Gpin}{Pinned genes}
}
\description{
Draw an interactive triwise dotplot in which genes can be interactively
}

