% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{interactiveDotplot}
\alias{interactiveDotplot}
\title{Interactive triwise dotplot}
\usage{
interactiveDotplot(Eoi, Gdiffexp = rownames(barycoords), Goi = c(),
  Glabels = rownames(Eoi), Gpin = c(), Coi = attr(barycoords,
  "conditions"), colorvalues = NULL, rmax = 5, sizevalues = c(T = 2, F =
  0.5), alphavalues = c(T = 0.8, F = 0.8), plotLocalEnrichment = F,
  width = NULL, height = NULL)
}
\arguments{
\item{Eoi}{Expression matrix with the three conditions in the columns}

\item{Gdiffexp}{Differentially expressed genes}

\item{Goi}{List with genes of interest}

\item{Glabels}{Labels for every gene if different from the rownames of `Eoi`}

\item{Gpin}{Pinned genes}
}
\description{
Draw an interactive triwise dotplot
}

