% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{interactivePvalplot}
\alias{interactivePvalplot}
\title{Interactive plot of p-values}
\usage{
interactivePvalplot(scores, gsetlabels, Coi, width = NULL, height = NULL)
}
\arguments{
\item{scores}{Dataframe as returned by `testUnidirectionality` containing for every gene set a q-value and an associated angle}

\item{gsetlabels}{Names of the gene sets, used for hovering}

\item{Coi}{Names of the three biological conditions, only used for labelling}

\item{width}{Width of the plot}

\item{height}{Height of the plot}
}
\description{
Draw a dotplot of p-values using a htmlwidget
}
\examples{
Eoi = matrix(rnorm(1000*3, sd=0.5), 1000, 3, dimnames=list(1:1000, c(1,2,3)))
Eoi[1:100,1] = Eoi[1:100,1] + 4 # the first 100 genes are more upregulated in the first condition
barycoords = transformBarycentric(Eoi)

gsets = list(a=1:50, b=80:150, c=200:500)
scores = testUnidirectionality(barycoords, gsets, Gdiffexp=(1:1000)[barycoords$r > 1])
\dontrun{
interactivePvalplot(scores, as.list(setNames(names(gsets), names(gsets))), 1:3)
}
}

