% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plotDotplot}
\alias{plotDotplot}
\title{Creating a dotplot}
\usage{
plotDotplot(barycoords, Gdiffexp = rownames(barycoords), Goi = NULL,
  Coi = attr(barycoords, "conditions"), colorby = "diffexp",
  colorvalues = NULL, rmax = 5, showlabels = T,
  sizevalues = setNames(c(0.5, 2), c(F, T)), alphavalues = c(T = 0.8, F =
  0.8), barycoords2 = NULL, baseangle = 0)
}
\arguments{
\item{barycoords}{Dataframe containing for every gene its barycentric coordinates, as return by \code{transformBarycentric}}

\item{Gdiffexp}{Differentially expressed genes}

\item{Goi}{Genes of interest, a character or numeric vector to plot one set of genes, a named list containing different such vectors to plot multiple gene sets}

\item{Coi}{Character vector specifying the names of the three biological conditions, used for labelling}

\item{colorby}{Color by differential expression ("diffexp") or by log fold-change ("z")}

\item{colorvalues}{Colors used according to colorby}

\item{rmax}{Number denoting the maximal radius of the grid. All points outside of the grid will be clipped on the boundaries.}

\item{sizevalues}{List with the size of each dot if differentially expressed (`T`) or not (`F`)}

\item{alphavalues}{List with the alpha value of each dot if differentially expressed or not}

\item{barycoords2}{Dataframe containing for every gene a second set of barycentric coordinates, as returned by \code{transformBarycentric}. An arrow will be drawn from the coordinates in `barycoords` to those in `barycoords2`.}
}
\value{
A ggplot2 object, which can be used to further customize the plot
}
\description{
Plot a dotplot
}
\examples{
data(vandelaar)
Eoi_replicates <- vandelaar[, phenoData(vandelaar)$celltype \%in\% c("BM_mono", "FL_mono", "YS_MF")]
Eoi <- avearrays(Eoi_replicates, phenoData(Eoi_replicates)$celltype)
Eoi = Eoi[,c("YS_MF", "FL_mono", "BM_mono")]

barycoords = transformBarycentric(Eoi)

}

