% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plotDotplot}
\alias{plotDotplot}
\title{Creating a dotplot}
\usage{
plotDotplot(E, barycoords = NULL, Gdiffexp = NULL, Goi = NULL,
  Coi = colnames(E), colorby = "diffexp")
}
\arguments{
\item{E}{Expression matrix}

\item{Gdiffexp}{Differentially expressed genes}

\item{Goi}{Genes of interest}

\item{colorby}{Color by differential expression ("diffexp") or by log fold-change ("z")}
}
\value{
Dataframe containing for every original point its new x and y coordinates in barycentric space
}
\description{
Plot a dotplot
}

