% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plotPvalplot}
\alias{plotPvalplot}
\title{Plot results from unidirectional enrichment}
\usage{
plotPvalplot(scores, Coi = c("", "", ""), colorby = NULL, showlabels = T,
  baseangle = 0)
}
\arguments{
\item{scores}{Dataframe as returned by `testUnidirectionality` containing for every gene set a q-value and an associated angle}

\item{Coi}{Names of the three biological conditions}

\item{colorby}{Column in scores used for coloring}

\item{showlabels}{Whether to show labels on the grid}

\item{baseangle}{The angle by which to rotate the whole plot (default to 0)}
}
\description{
Plot results from unidirectional enrichment
}

