% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plotPvalplot}
\alias{plotPvalplot}
\title{Plot results from unidirectional enrichment}
\usage{
plotPvalplot(scores, Coi = c("", "", ""), colorby = NULL,
  showlabels = TRUE, baseangle = 0)
}
\arguments{
\item{scores}{Dataframe as returned by `testUnidirectionality` containing for every gene set a q-value and an associated angle}

\item{Coi}{Names of the three biological conditions, only used for labelling}

\item{colorby}{Column in `scores` used for coloring}

\item{showlabels}{Whether to show labels on the grid}

\item{baseangle}{The angle by which to rotate the whole plot (default to `0`)}
}
\value{
A ggplot2 plot, which can be used for further customization
}
\description{
Plots each enriched gene set as a dot on a dotplot
}
\examples{
Eoi = matrix(rnorm(1000*3, sd=0.5), 1000, 3, dimnames=list(1:1000, c(1,2,3)))
Eoi[1:100,1] = Eoi[1:100,1] + 4 # the first 100 genes are more upregulated in the first condition
barycoords = transformBarycentric(Eoi)

gsets = list(a=1:50, b=80:150, c=200:500)
scores = testUnidirectionality(barycoords, gsets, Gdiffexp=(1:1000)[barycoords$r > 1])

plotPvalplot(scores)
}

