% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plotRoseplot}
\alias{plotRoseplot}
\title{plotRoseplot}
\usage{
plotRoseplot(barycoords, Gdiffexp, Goi = rownames(barycoords),
  conditions = attr(barycoords, "conditions"), nbins = 12,
  bincolors = rainbow(nbins, start = 0, v = 0.8, s = 0.6))
}
\description{
A rose plot shows the distribution of the differentially expressed genes in different directions.
}

