% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{testEnrichment}
\alias{testEnrichment}
\title{Simple fisher's exact test for enrichment in a given set of genes of interest (Goi)}
\usage{
testEnrichment(Goi, gsets, background, minknown = 2, mindiffexp = 2,
  maxknown = 500)
}
\arguments{
\item{Goi}{character vector containing the genes tested for enrichment}

\item{gsets}{List of character vectors, each containing a set of genes (gene identifiers)}

\item{background}{character vector containing background gene identifiers}

\item{minknown}{Minimal number of genes within a gene set for it to be considered for enrichment}

\item{mindiffexp}{Minimal number of genes differentially expressed within a gene set for it to be considered for enrichment}

\item{maxknown}{Maximal number of genes within a gene set for it to be considered for enrichment}
}
\value{
Dataframe containing for every gene set which passed the filtering: \itemize{
 \item p-value of enrichment
 \item q-value of enrichment (p-value corrected for multiple testing)
 \item estimated odds score of enrichment, how much more likely is a given gene set to be part of
}
}
\description{
Simple fisher's exact test for enrichment in a given set of genes of interest (Goi)
}
\examples{
Goi = 1:50
gsets = list(a= 20:70, b = 45:80)
background = 1:100
testEnrichment(Goi, gsets, background)
}
