% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{testLocality}
\alias{testLocality}
\title{Test local upregulation}
\usage{
testLocality(Goi, Gdiffexp, angles, deltangle = pi/24,
  bandwidth = pi/3)
}
\arguments{
\item{Goi}{Gene set for which to test enrichment}

\item{Gdiffexp}{Differentially expressed genes}

\item{angles}{Numeric vector with angles of all genes or a dataframe as returned by `transformBarycentric`}

\item{deltangle}{Stepsize of angles}

\item{bandwidth}{Bandwidth of angles}
}
\value{
Corrected p-values for enrichment around every angle
}
\description{
Tests for local upregulation locally in certain directions
}
\examples{
Eoi = matrix(rnorm(1000*3, sd=0.5), 1000, 3, dimnames=list(1:1000, c(1,2,3)))
Eoi[1:100,1] = Eoi[1:100,1] + 4 # the first 100 genes are more upregulated in the first condition
barycoords = transformBarycentric(Eoi)

Goi = 1:50
qvals = testLocality(Goi, Gdiffexp=(1:1000)[barycoords$r > 1], barycoords)
plot(attr(qvals, "anglesoi"), qvals)
}
