% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{testRayleigh}
\alias{testRayleigh}
\title{Rayleigh z-test implementation}
\usage{
testRayleigh(angles)
}
\arguments{
\item{angles}{Numeric vector containing angles in radians}
}
\value{
P-value of unidirectionality under the uniformity null hypothesis
}
\description{
Rayleigh z-test implementation
}
\examples{
testRayleigh(as.numeric(circular::rvonmises(10, circular::circular(1), 2)))
testRayleigh(seq(0, 2*pi, 0.1))
}

