% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R, R/analyze2.R
\name{testUnidirectionality}
\alias{testUnidirectionality}
\title{Test gene sets for unidirectional enrichment}
\usage{
testUnidirectionality(angles, gsets, Gdiffexp = NULL, minknown = 2,
  minfound = 2, maxknown = 500, weight = T, angleweights = NULL)

testUnidirectionality(angles, gsets, Gdiffexp = NULL, minknown = 2,
  minfound = 2, maxknown = 500, weight = T, angleweights = NULL)
}
\arguments{
\item{angles}{either a dataframe obtained from transformBarycentric or a list of angles with gene ids as names}

\item{Gdiffexp}{differentially expressed genes}

\item{angles}{either a dataframe obtained from transformBarycentric or a list of angles with gene ids as names}

\item{Gdiffexp}{differentially expressed genes}
}
\value{
Dataframe containing for every gene set which passed the filtering: \itemize{
 \item p-value of unidirectionality
 \item q-value of unidirectionality (p-value corrected for multiple testing)
 \item average angle
}

Dataframe containing for every gene set which passed the filtering: \itemize{
 \item p-value of unidirectionality
 \item q-value of unidirectionality (p-value corrected for multiple testing)
 \item average angle
}
}
\description{
Test gene sets for unidirectional enrichment

Test gene sets for unidirectional enrichment
}

