% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{testUnidirectionality}
\alias{testUnidirectionality}
\title{Test gene sets for unidirectional enrichment}
\usage{
testUnidirectionality(barycoords, gsets, bm = NULL, minknown = 5,
  minfound = 5, maxknown = 500, mc.cores = options("mc.cores"))
}
\arguments{
\item{gsets}{list of character vectors, each containing a set of genes (gene identifiers)}

\item{angles}{either a dataframe obtained from transformBarycentric or a list of angles with gene identifiers as names}

\item{Gdiffexp}{differentially expressed genes}
}
\value{
Dataframe containing for every gene set which passed the filtering: \itemize{
 \item p-value of unidirectionality
 \item q-value of unidirectionality (p-value corrected for multiple testing)
 \item average angle
}
}
\description{
Test gene sets for unidirectional enrichment
}

