% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{transformBarycentric}
\alias{transformBarycentric}
\title{Conversion to barycentric coordinates}
\usage{
transformBarycentric(E, transfomatrix = NULL)
}
\arguments{
\item{E}{Expression matrix}
}
\value{
Dataframe containing for every original point its new coordinates in `d-1` dimensions
}
\description{
Converts the expression matrix containing two or more biological conditions to barycentric coordinates, reducing its dimensionality by one while retaining information of differential expression.
}

