% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R, R/analyze2.R
\name{transformBarycentric}
\alias{transformBarycentric}
\title{Conversion to barycentric coordinates}
\usage{
transformBarycentric(E, transfomatrix = NULL)

transformBarycentric(E, transfomatrix = NULL)
}
\arguments{
\item{E}{Expression matrix}

\item{transfomatrix}{Transformation matrix}

\item{E}{Expression matrix}

\item{transfomatrix}{Transformation matrix}
}
\value{
Dataframe containing for every original point its new x and y coordinates in barycentric space

Dataframe containing for every original point its new x and y coordinates in barycentric space
}
\description{
Converts the expression matrix containing 3 biological conditions to barycentric coordinates.

Converts the expression matrix containing 3 biological conditions to barycentric coordinates.
}

