% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{transformBarycentric}
\alias{transformBarycentric}
\title{Conversion to barycentric coordinates}
\usage{
transformBarycentric(E, transfomatrix = NULL)
}
\arguments{
\item{E}{Expression matrix}

\item{transfomatrix}{\code{NULL} (default) or a numeric matrix containing the transformation matrix}
}
\value{
Dataframe containing for every original point its new coordinates in `d-1` dimensions
}
\description{
Converts the expression matrix containing three biological conditions to barycentric coordinates, reducing its dimensionality by one while retaining information of differential expression.
}
\examples{
Eoi = matrix(rnorm(1000*3, sd=0.5), 1000, 3, dimnames=list(1:1000, c(1,2,3)))
Eoi[1:100,1] = Eoi[1:100,1] + 1
barycoords = transformBarycentric(Eoi)
hist(barycoords$angle)
}

