% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{transformReverseBarycentric}
\alias{transformReverseBarycentric}
\title{Convert from barycentric coordinates back to original expression values}
\usage{
transformReverseBarycentric(barycoords, transfomatrix = NULL)
}
\arguments{
\item{barycoords}{Dataframe of barycentric coordinates with x and y in separate columns}

\item{transfomatrix}{\code{NULL} (default) or a numeric matrix containing the transformation matrix}
}
\value{
Matrix containing for every gene original expression values centered around 0
}
\description{
Converts a dataframe contain barycentric coordinates in the x and y columns back to original coordinates (apart from a constant for each gene).
}
\examples{
Eoi = matrix(rnorm(1000*3, sd=0.5), 1000, 3, dimnames=list(1:1000, c(1,2,3)))
Eoi[1:100,1] = Eoi[1:100,1] + 1
barycoords = transformBarycentric(Eoi)
reversebarycoords = transformReverseBarycentric(barycoords)
all.equal(scale(t(Eoi)), scale(t(reversebarycoords)), check.attributes=FALSE)
}

