% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vandelaar}
\alias{vandelaar}
\title{Expression dataset from van de Laar et al. 2016}
\format{An object of class \code{ExpressionSet} with 20917 rows and 36 columns.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE76999}
}
\usage{
vandelaar
}
\description{
This dataset contains, among others, gene expression data from three macrophage progenitor populations
}
\keyword{datasets}

